requirements.txt
jdatetime==3.1.0
psycopg2-binary==2.7.5
python-telegram-bot==11.1.0
SQLAlchemy==1.2.15
Unidecode==1.1.1
loguru
persian
pandas
pytest
xlrd
xlsxwriter